+++ src/common/js/constants.js
@@ -2,18 +2,37 @@
  * pwix:layout/src/common/js/constants.js
  */
 
-// verbosity levels
+Layout.C = {
+    // the size of the breakpoints as defined in /src/client/stylesheets/layout.less
+    Breakpoints: {
+        XS:  384,
+        SM:  432,
+        MD:  800,
+        LG: 1024,
+        XL: 1920
+    },
 
-UI_VERBOSE_NONE = 0;
-UI_VERBOSE_CONFIGURE = 0x01 << 0;
+    // verbosity levels
+    Verbose: {
+        NONE: 0,
+        CONFIGURE: 0x01 << 0
+    },
 
-// the size of the breakpoints as defined in /src/client/css/ui_constants.less
-// the size of the breakpoints as defined in /src/client/css/ui_layout.scss
-UI_XS_WIDTH =  384;
-UI_SM_WIDTH =  432;
-UI_MD_WIDTH =  800;
-UI_ST_WIDTH = 1024;
-UI_LG_WIDTH = 1920;
+    // the view model
+    View: {
+        XS: 'XS',
+        SM: 'SM',
+        MD: 'MD',
+        LG: 'LG',
+        XL: 'XL'
+    }
+};
+
+Layout.C.Breakpoints.XS =  384;
+Layout.C.Breakpoints.SM =  432;
+Layout.C.Breakpoints.MD =  800;
+Layout.C.Breakpoints.ST = 1024;
+Layout.C.Breakpoints.LG = 1920;
 
 // the view model
 UI_VIEW_XS = 'UI_VIEW_XS';
@@ -24,11 +43,11 @@ UI_VIEW_LG = 'UI_VIEW_LG';
 UI_VIEW_XL = 'UI_VIEW_XL';
 
 Layout.iBreakpoints = [
-    UI_XS_WIDTH,
-    UI_SM_WIDTH,
-    UI_MD_WIDTH,
-    UI_ST_WIDTH,
-    UI_LG_WIDTH
+    Layout.C.Breakpoints.XS,
+    Layout.C.Breakpoints.SM,
+    Layout.C.Breakpoints.MD,
+    Layout.C.Breakpoints.ST,
+    Layout.C.Breakpoints.LG
 ];
 
 Layout.cBreakpoints = [