+++ src/common/js/index.js
@@ -3,7 +3,5 @@
  */
 
 import './global.js';
-//
 import './constants.js';
-//
 import './configure.js';